requirements.txt
gradio
python-dotenv
openai
youtube_transcript_api